de-dupe variable setup and cleanup vars
This patch externalizes the variable setup so all clients that will be
setup can re-use the variable setup method (openstack_variables.get()).
Also in this patch is a cleanup of variables that are no longer needed,
ensuring the ther users interactive environment is as sterile as we can
make it.